[Support] Use the Modular OID prefix. (#23650)
We now have a number from IANA, so use that as the canonical modular OID prefix.

MODULAR_ORIG_COMMIT_REV_ID: 71a046420f3ca7a67ddeb467c11d70022fcaf66b

+++ Support/lib/ASN1/ObjectID.cpp
@@ -219,8 +219,8 @@ llvm::raw_ostream &M::operator<<(llvm::raw_ostream &os,
                                  const ASN1::ObjectID &oid) {
   if (oid.isModularOID()) {
     llvm::interleave(
-        ASN1::ObjectID::getModularPrefix(), os,
-        [&](uint8_t u) { os << (int)u; }, ".");
+        ASN1::ObjectID::getModularPrefix(), os, [&](uint64_t u) { os << u; },
+        ".");
 
     // Separator dot between prefix and the arc.
     os << ".";
